SVM: add `program_loader` module (#1017)
* SVM: make `load_program_accounts` a helper

* SVM: make `load_program_from_bytes` a helper

* SVM: make `load_program_with_pubkey` an associated function

* SVM: make `load_program_with_pubkey` a helper

* SVM: add `program_loader` module

* SVM: update spec for `program_loader`
